API hanged on a termination signal
When murano-api is running with several api worker threads sending it a
termination signal (SIGTERM, SIGHUP or SIGINT) leads to waitall()
being called inside a GreenPool's greenthread. The reason is that a
wsgi server is started in a green thread from the same green pool that
is passed to the server itself to be used for spawning client green
threads. To avoid it, it is reasonable to use different pools for
spawning a wsgi server and for its internal usage.

Closes-Bug: #1536666
Change-Id: Ib8c9609035a5585b0392811ebd9de669dbc0c986