Basic reflection capabilities were added to MuranoPL
Now it is possible to get type info with typeinfo() function and using
it as a starting point obtain information about the class, its methods
and properties as well as the package of the class.
See TestReflection.yaml for detailed usage.

Also some refactoring to existing code model were made.

Change-Id: Ic8a719e874126d735c26b956a49cb790d434c64c