Strengthen validation of receiver type for methods
When static method was called on a type object
(ns:Type.method()) type object was used only to load
right context but wasn't actually validated. As a result
if the caller tried to call static method that doesn't
exist in the target class but does exist in the caller's
class it was called instead of exception being raised.

This commits improves MuranoTypeParameter
by providing ability to specify optional base class
for requested type object and then uses this functionality
to check receiver of the call expression (type object)

Change-Id: Iff1c7c4b08e04cbddaa3ff0029a8cfa0a8360ce5