Makes config() function be available to Core Library only
config() function returns information from murano.conf.
But since it contains sensitive information like passwords
it creates a security issue for Murano because any app can
get access to that information. However the function is
used by the core library to get RabbitMQ credentials.

This commit makes the function available to Core Library
only which currently identified by the name (io.murano)
but in the future will be identified by the package signature.

Change-Id: I3fe5c153f931decc59bc8bf9eb87c78d459a64fa
Closes-Bug: #1506807