Removes ability to alter defaults for child objects
Previously there was a feature which allowed to set
new default property values for owned objects
when using class() contract. This was done using
Default property keyword which served for 2 different
purposes in this case. The feature was never used neither
documented anywhere. And it wasn't ever really designed.
It also had a bug that caused default value [] (empty list)
to be replaced with {} (empty dict).

But there are cases when instead of $.class() contract
we have [$class()] contract and in this case default
value should also be a list rather then a dictionary of
property values. So the only reason some applications
worked until recently was that bug.

However recent refactoring reworked that code and
fixed that bug. So, in turn, it broke all docker applications
that implicitly relied on that bug.

This commit removed the entire feature. It was never
used thus will break. We can reconsider it once we have
a better design and use cases for it.

Change-Id: I792fbd9801ff7776ee72660b2be2f89d0bfeb132
Closes-Bug: #1558332