Handling of ephemeral methods was fixed
Ephemeral method is the one that created
on the fly and doesn't belong to extension
class. Such methods were returned as a replacement
for ambiguous methods. Such ephemeral method
would throw exception upon attempt to invoke it so that
we get this exception on run-time rather that load-time.
However yaql FunctionDefinition code wasn't ready for
such functions

Change-Id: Ica8585b9750277d29f571b90dc7b92e12e5fef30
Closes-Bug: #1562011