Don't use bash eval for Linux.runCommand
Linux.runCommand method used to pass script
in a command line argument and evaluated it using
"eval". This approach failed when the script contained
input redirection.

With this fix script is passed as a file which is specified
as an entry point in execution plan.
Also we supported explicit files in execution plan but not
for the script entry point. This was fixed as well.

Change-Id: Ifd15fc57b08e33ed61cd2765613cd705ce12a5f2
Closes-Bug: #1562954