Track status report timestamp
Report status messages were ordered by the time
they were inserted into database. And that is the
time they were received and processed not the
time the report message was generated. So if
deployment generates reports faster than they
get processed it can be that they accumulate
in RabbitMQ and then received by the API not
in original order. Another possibility is that
several API instances would get those messages
and because of the race condition insert them
into database in wrong order.

With this fix report timestamp becomes part
of the report itself and is inserted into "created"
table column. Thus created is now when the
report was created, not the database record.

Change-Id: I5b705ee95c6fa2275937513aba0bcb0053c0901c
Closes-Bug: #1462270