Fixes race condition in HeatStack
Before each operation HeatStack waits for the stack to become available.
But if several operations happen concurrently it might be the case when
several threads both check stack availability and conclude its ready and
then both try to update it. In this case Heat client throws HTTPConflict
exception. This fix handles it and retries the operation after delay.

Change-Id: Ibff461ed606e7fb1fc76ad422b4431476e22ab0b
Closes-Bug: #1571863