Incorrect Method param inheritance was fixed
When a class calls method that was overridden in a child
class in yaql context both implementation are present. So if
derived implementation happen to have another signature
normal yaql resolution rule will apply and both of them
can be called depending on the parameters. However
this is goes against MuranoPL design that prohibits
several versions of the same method to be accessible in
class hierarchy. With this fix derived implementation
will always signature with *args/**kwargs so it will
always match and take precedence over that from the base class

Closes-bug: #1572215
Change-Id: I0ce27d03d785675f166fd265bb30b3b70ca6c7cb