Fixed a bug in _get_glare_url
Due to a typo in the _getGlare_url method Murano was not able to
properly retrieve the Glare endpoint url from the Keystone's service
catalog.
This lead to an HTTP 400 when the package UI was requested from
murano-api if the murano.config did not contain a pre-configured Glare
URL.
This has been addresed.

Change-Id: I6447ee0cc3a322c841f5e6d27c74ef39cad1d063
Closes-Bug: #1572518