Revert the destroy execution order
Restores back the order of .destroy execution so that
child objects get destroyed first and Environment get
destroyed last.

Both orders have their cons and pros. With
the current order Environment and thus Heat stack
are delete first so if the app sends a command to
murano-agent in its destroy method it will never
receive an answer since the VM of the agent was
already terminated. This broke Kubernetes and might
broke other apps as well.  However the downside is that
apps are going to perform uninstall of the software
on VM that is going to be terminated anyway which
takes time. But until we implement a controllable
resource deallocation this looks like a safer approach.

Change-Id: I8265a69938658e89e4fc36824502c75a75a75c66
Closes-Bug: #1576703