Documentation to use sub-templates in HOT packages
The document provides the description of adding sub-templates in Murano
HOT packages. It is possible to do a template composition by referencing
some existing template by putting it file name into a "type" attribute
of resource definition as if it was some real resource. To enable this
use case sub-templates need to be placed in Resources/HotFiles.

Change-Id: I1ce0d264d02456645cf59067075a980556ce0e98
Closes-Bug: #1579813