Add metadata to the result of action serialization
In addition to action name, now title, description and help text
also appear in the serialized action, if corresponding metaclasses
are attached to the method. If no title specified by metadata, it
is equal to method name in serialization result.

Tests are updated to check new functionality. All test cases for
actions serialization are now gathered in one place.

Change-Id: I00c46677ed6bef3c8fcf93f07a1abc004c7304d3
Partially-implements: blueprint static-actions