Two-phase Instance deployment support
Currently the provisioning of VM Instances is done by calls of the
'deploy' method of 'io.murano.resources.Instance' class. This method
generates the snippet of Heat template describing the VM and its
resources, merges this snippet into the global template of the
environment and then pushes this template to Heat API.

Thus the provisioning of N instances can currently be done only with N
updates of the Heat stack, i.e. with N calls of the Heat API, each
followed by a significant wait interval.
The proper way to do this would be to generate a single Snippet for
all the VMs at once, then execute a single Heat Stack update - and then
process its output also for all the VMs at once.

To allow application developers to implement such a behavior, the
'deploy' method of Instance class is split into two parts:
generation of the heat template snippet and the actual pushing of the
snippet into the Heat API.
These two phases are implemented as 'beginDeploy' and 'endDeploy'
methods of the Instance class. These methods are supposed to be
properly called by the application code willing to provision multiple
instances at once, while the old 'deploy' method just calls them
sequentially and is left for backwards compatibility.

Similar approach is implemented for releaseResources routine.

Change-Id: I985883816bbe9dcc23b05a4170d22acd4af9222d
Targets-blueprint: scalable-application-framework