Added a muranoPL-specific override of 'call'
Yaql's standard system function 'call' was not working properly in
MuranoPL because murano relies on object-specific context containing
definitions of the methods of appropriate class.

This patch adds a MuranoPL-specific override of this function which
creates an appropriate method-invocation AST and executes it. This
will invoke regular muranoPl's override of operator_. thus causing all
the appropriate type and object contexts to be created for the call.

Closes-bug: #1595174

Change-Id: If3e98a60063b0843c317fa565616c8d88f9e497c