Replication classes of Application Development Framework library
This is a first commit for a library package for Application
Development Library (earlier known as "Scalable Framework") called
io.murano.applications

This patch includes the base classes for node Replication
(ReplicaProvider and ReplicationGroup) and their Test-Runner-based
tests, as well as two reference implementations of ReplicaProvider -
one based on the templating (TemplateReplicaProvider) and another one
based on the cloning of passed objects (CloneReplicaProvider).

Both reference provider implementations are covered with their own
tests.

Targets-blueprint: application-development-framework
Change-Id: Ia6a48d51f41693fbb85d2b6c2851c1c286c6aa23