Use upper-constraints in tox test environments
This will force pip install to use the upper-constraints.txt specified
version of pip modules. When you don't do this, you are out on the
bleeding edge and become unstable everytime some python library in the
wide world changes in a way that you don't expect.

The releasenotes job has no constraints file, so override the install
command there.

This commit is being proposed on stable/liberty and therefore the
install command includes the clear designation of the branch. This is
because, on branches other than master, we need to be sure to get the
upper-constraints.txt file for that branch.

Change-Id: Ib08baa8f57a281c75247ba3d9b9ab0211fe5d4fd
Partial-bug: #1598906
(cherry picked from commit 317e9bfe5e50b26189f0e098658dc41d689204bd)