Fix used package versions for stable inheritance
When an object from an object model is deserialized
on subsequent deployment it's version is usually
fixed in the object model. However the version of
base classes for that object is not as long as it meets
to the class requirements. Thus slightly different version
might be used on each deployment that does satisfies the
requirements but may have different internal state representation.

To solve this a list of all package versions that are in use by the
object model is written to the hidden section of the object model.
On the subsequent load package loader will prefer versions in the
list over newer versions available in the catalog at that time

Change-Id: I17087c0045210d28b87aba4d50c329f688df3954
Closes-Bug: #1602839