Ability to instantiate object graph
Adds ability to load object sub-graph in a single call.
The graph may either be defined in the serializable
object-model format

?:
  id: ...
  type: ...
  classVersion: ...
  package: ...
 prop1: value
 prop2: value2

 or in the MuranoPL format

 ns:Type:
   prop1: value1
   prop2: value2
 id: ...
 name: ...

During the load IDs of the loaded objects are re-generated so
that explicit specification of IDs is only necessary for specification
of references within the graph and are not retained.

The load is done in the temporary ObjectStore. Loaded objects are
copied to the main ObjectStore then.

3rd overload of the new() function was added to deserialize the model:
new($model, $owner)

Change-Id: I849a5d6c9a28843398a4c4c884a1c3138ac7ede3