An initial commit for MuranoPL garbage collector
This is an initial commit for MuranoPL garbage collector implementation.
A first step of GC implementation is implementation of destruction
dependencies between applications. Destruction dependencies works as
described below.
Since murano still have a problem described in a spec, it was decided
to skip serialization/deserialization steps for destruction dependencies
for now at least.
If object Foo establishes such a dependency on the object Bar then:
* Foo will be notified when Bar is about to be destroyed.
* If both Foo and Bar are going to be destroyed in the same garbage
  collection execution, Bar will be destroyed before Foo.

Change-Id: I33eb8e09773aebe0a08542b70171466a0b849885
Targets-blueprint: dependency-driven-resource-deallocation