Nested new() were not using 2-phase load
Usually object model load happens in two phases. This is also true for
new($model) that does the same at run time. However if the new()
happens to be in .init method that is executed as part of object model
itself it becomes nested model load. However due to the bug instead of
starting another 2-phase session, new() used to join the same session
.init run in (second pass of the outer load). Thus cross object
references within the nested model were not resolved.

Closes-Bug: #1608718
Change-Id: I81396291c3beaee8b0c5b50879313cc236a5aa77