Move .init call into separate load phase
Before this commit .init method were called as the last step
of second phase initialization. This is also the phase when
forward references are get resolved.

Because load phases work recursively second phase of load
for the nested objects will happen before the outer counterparts.
If those nested objects were to access their owners in the .init
method it can cause exception to be thrown because forward
reference of the owner were not yet resolved.

This commits moves .init calls out of the second phase. Now
during the second phase we collect everything that need to
be called and call it after the complete graph load.
This also guarantees that we call .init only on the object graph
that can be loaded (since it already loaded by that time)

Change-Id: Ie86655bcc668be880a7b97695835ccddb9adda0f