Don't put non-initialized objects into ObjectStore
Because object model deserialization happens in
several phases nested objects without fixed ID
will be created again upon each pass and only
the last one will be the property value. Previously
created objects remain in the ObjectStore but were
not initialized. This is going to cause issues once garbage
collection is implemented because it may attempt to
collect uninitialized objects.

With this commit objects that are not fully initialized are
not copied into the permanent object store anymore.

Change-Id: I6924b93d79c69181fe2cc8baa329a3822892af3c