Changes type format in object model
Since introduction of the statics into MuranoPL
type object (ns:Class) becomes available to the
MuranoPL code as a data that can be assigned to
the property (argument, variable). But type object
is not JSON serializable. Thus if the object model
contains type object (for example it could be the
dictionary returned by the dump() function) it
cannot be serialized any longer.

If we serialize type object into dictionary with
three attributes like we do for MuranoObject (type,
classVersion, package) this will not be compatible
with existing applications in case of dump()
outcome because for MuranoObjects those
three attributes were located right in the '?' section
rather than in a nested dictionary.

The solution is to encode all those three attributes into
a single string and teach serializer to produce such string
each time it sees type object in the object model.

This commit changes type string format into
"typeName/version@package"  where version and package
parts are optional. When deserializing object model loader
 still uses current "className" and "package" attributes as
 a defaults for missing parts of the new type string so
 backward compatibility is retained except for the case where
 the client put short type string into the type attribute of the
 "?" and don't expect it to change after the deployment.

Change-Id: Icd3b5e238ad09584b4b7fdfb74120820078715c6