Renamed Instance* to Server*
The naming of the classes which were working with servers was
inconsistent: at some places they were referenced as "instances", at
some others as "servers". This is inconsistent, also the word
'instance' is too generic.

So a refactoring was made: InstanceGroup and InstanceProvider were
renamed to ServerGroup and ServerProvider. Some other properties and
variable names were renamed accordingly.

Change-Id: I4ce2d09ae890aec4eb61d9719932c65457fb6dba