TestFixture mocks were updated
Core library was updated as part of the multi-region support
effort. FormatVersion of the core library was incremented and,
as a result $dict.key expressions started to work as they are in
yaql 1.0 rather than in 0.2 and the presence of the key become
mandatory. However Network mock in the TextFixture used to
return empty dictionary and it caused Instance class to fail.
The commit adds missing keys to the dictionary returned by
the joinInstance() method.

Also the test runner was improved to print the stack trace
upon failures so that it is possible to identify where the
error happened.

Change-Id: Ia7b0f61370effe4f13fed4ec85e806f3b0fdb80b