ObjectStore parent lookup with temporary id was fixed
When an existing object is used as a template()
contract source it is dumped and then recreated
from the serialized model. Serialized object IDs
are considered to be temporary so that the result
object get different ID from the clone source.
But since merge of Objects and ObjectsCopy
object model sections ObjectStore started accessing
parent store on attempt to get the object by its ID.
And since the source object ID and temporary ID
are equal it returned the source object rather than
the clone

Targets-blueprint: dependency-driven-resource-deallocation

Change-Id: I1bc3fef85e1fb91b139bd916fdaa1ebbc6d08a19