Add __ne__ built-in function for murano
In Python 3 __ne__ by default delegates to __eq__ and inverts the result,
but in Python 2 they urge you to define __ne__ when you define __eq__ for
it to work properly [1].There are no implied relationships among the
comparison operators. The truth of x==y does not imply that x!=y is false.
Accordingly, when defining __eq__(), one should also define __ne__() so that
the operators will behave as expected.

Change-Id: If58b0a63004a040b9639aa1315716973c45d13f2