Raise exception on call of method of destroyed object
In theory using destruction dependencies one can obtain a new
reference to an object that is about to be destroyed. In this
case any attempt to call a method on such object should result
in exception being raised. This guarantees that the nothing can
be done on the object after its .destroy method was executed.

Also because the HeatStack instance is owned by the CloudRegion
it is going to be destroyed prior to it. But CloudRegion needs
to call $stack.delete() in its .destroy method and this will cause
error. As a solution HeatStack made not to be owned by the
region and instead be an independent object with destruction
dependency on it which guarantees that it will be destroyed
after the CloudRegion.

Targets-blueprint: dependency-driven-resource-deallocation
Change-Id: I5650ea672d5f121be69669f27dd5f513fbdd5c44